backend/app/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.7.0
django-mysql==3.11.1
djangorestframework==3.12.2
mysqlclient==2.0.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
pyJWT==2.0.1
Faker==6.6.1
django-redis==4.12.1
stripe==2.56.0